+++ src/com/company/SocketIOHandler.java
@@ -1,4 +1,4 @@
-package com.company.Client;
+package com.company;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,21 +6,20 @@ import java.io.InputStreamReader;
 import java.io.PrintWriter;
 import java.net.Socket;
 
-public class ClientIOHandler {
+public class IOHandler {
+    private PrintWriter output;
+    private BufferedReader input;
 
-    private final PrintWriter output;
-    private final BufferedReader input;
-
-    public ClientIOHandler(Socket clientSocket) throws IOException {
+    public IOHandler(Socket clientSocket) throws IOException {
         this.output = new PrintWriter(clientSocket.getOutputStream(), true);
         this.input = new BufferedReader(new InputStreamReader(clientSocket.getInputStream()));
     }
 
-    public String readFromSocket() throws IOException {
-        return this.input.readLine();
-    }
-
     public void printToSocket(String inputLine) {
         this.output.println(inputLine);
+    }
+
+    public String readFromSocket() throws IOException {
+        return this.input.readLine();
     }
 }
